moved the user dto to their designated directory
