--- src/main/java/ba/edu/ibu/finance_tracker/rest/DTO/UserSearchResultDTO.java
@@ -1,48 +1,0 @@
-package ba.edu.ibu.finance_tracker.rest.dto;
-
-public class UserSearchResultDTO {
-    private String id;
-    private String name;
-    private String surname;
-    private String email;
-
-    public UserSearchResultDTO(String id, String name, String surname, String email) {
-        this.id = id;
-        this.name = name;
-        this.surname = surname;
-        this.email = email;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getSurname() {
-        return surname;
-    }
-
-    public void setSurname(String surname) {
-        this.surname = surname;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-}
