+++ src/main/java/ba/edu/ibu/finance_tracker/rest/dto/UserDTO/UserCreateRequestDTO.java
@@ -1,0 +1,60 @@
+package ba.edu.ibu.finance_tracker.rest.dto.UserDTO;
+
+public class UserCreateRequestDTO {
+
+    private String name;
+    private String surname;
+    private String email;
+    private String password;
+    private double balance;
+    private String parentId;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getSurname() {
+        return surname;
+    }
+
+    public void setSurname(String surname) {
+        this.surname = surname;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public double getBalance() {
+        return balance;
+    }
+
+    public void setBalance(double balance) {
+        this.balance = balance;
+    }
+
+    public String getParentId() {
+        return parentId;
+    }
+
+    public void setParentId(String parentId) {
+        this.parentId = parentId;
+    }
+
+}
